+++ com.unity.mobile.notifications/Runtime/Android/Plugins/com/unity/androidnotifications/UnityNotificationManager.java
@@ -577,6 +577,11 @@ public class UnityNotificationManager extends BroadcastReceiver {
         return alarmManager.canScheduleExactAlarms();
     }
 
+    public boolean canScheduleExactAlarms() {
+        AlarmManager alarmManager = (AlarmManager) mContext.getSystemService(Context.ALARM_SERVICE);
+        return canScheduleExactAlarms(alarmManager);
+    }
+
     // Call AlarmManager to set the broadcast intent with fire time and interval.
     private void scheduleNotificationIntentAlarm(long repeatInterval, long fireTime, PendingIntent broadcast) {
         AlarmManager alarmManager = (AlarmManager) mContext.getSystemService(Context.ALARM_SERVICE);