+++ com.unity.mobile.notifications/CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this package will be documented in this file.
 - [Android] - Added support for notification channel groups.
 - [Android] - Added support for BigPicture style.
 - [Android] - Icons now support local files and URIs.
+- [Android] - Added API to check if UI with permission rationale for notifications should be shown before requesting permission.
+- [Android] - Added APIs to request exact scheduling and bupasssing battery optimizations.
 - [iOS] - Added support for notification action icons (requires iOS 15).
 - [iOS] - Added property to set interruption level (requires iOS 15).
 - [iOS] - Added property to set relevance score (requires iOS 15).